requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
decorator==5.1.1
Django==4.0.2
django-3-jet==1.0.8
django-captcha-admin==0.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-gravatar2==1.4.4
django-jalali==5.1.0
django-js-asset==2.0.0
django-multi-captcha-admin==2.0.0
django-recaptcha==1.3.0
django-recaptcha2==1.4.1
django-tinymce==3.4.0
executing==0.8.2
feedparser==6.0.8
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.4
idna==3.3
ipython==8.0.1
ipython-genutils==0.2.0
jdatetime==3.8.2
jedi==0.18.1
Khayyam==3.0.17
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauth2client==4.1.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.27
protobuf==3.19.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
